odhcpd: remove the "legacy" option
LuCI has already been updated so that the "dhcpv4" option is exposed in the UI
and the option also gets set when the user clicks the "Enabled DHCP" option in
the LuCI UI. Furthermore, there are instructions in the UI to enable the DHCPv4
server on a per-interface basis.

Finally, the odhcp defaults script (in the openwrt repo) also sets the dhcpv4
option, as necessary.

People who don't use LuCI but prefer the cmdline should presumably be able to
figure out that the dhcpv4 option needs to be set (if it isn't already), so
remove the "legacy" option.

This supersedes this PR:
https://github.com/openwrt/odhcpd/pull/275

Note that the "maindhcp" option is used by both dnsmasq and odhcpd to control
which of them handles DHCPv4, and is now exposed in the LuCI UI as well, so it
cannot be removed.

Signed-off-by: David Härdeman <david@hardeman.nu>