dhcpv4: update ubus DHCPv4 events/methods
This is based on the assumption that we don't really have any consumers
of ubus DHCPv4 events (yet).

With that in mind, rename and simplify the event function (yes, we
should add a sibling function for DHCPv6 leases later).

Also, take the chance to improve some naming, and introduce additional
attributes to the ubus event/method (the interface name and the
DUID/IAID, if known).

Signed-off-by: David Härdeman <david@hardeman.nu>
Link: https://github.com/openwrt/odhcpd/pull/306
Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>