router: only store PIO prefix and length
Right now, router_add_ra_pio() stores the whole address of the interface
in iface->pios, together with the prefix_len assigned to the interface
rather than the actual information which is announced in RA PIOs.

This can also be seen by reading the JSON file, e.g. (using a veth pair
where odhcpd listens to foo-server and we run commands on foo-client):

	$ ip addr add fd00:bbbb::1/48 dev foo-server
	$ start odhcpd
	$ ip -6 addr show dev foo-server
	...
	    inet6 fd00:bbbb::1/48 scope global
	       valid_lft forever preferred_lft forever
	...

	$ rdisc6 foo-client
	...
	 Prefix                   : fd00:bbbb::/64
	  On-link                 :          Yes
	  Autonomous address conf.:          Yes
	  Valid time              :         5400 (0x00001518) seconds
	  Pref. time              :         2700 (0x00000a8c) seconds
	...

	$ cat /tmp/odhcpd-prefixes/odhcpd.pio.foo-server
	...{"prefix":"fd00:bbbb::1","length":48}...

This leads to warnings being printed e.g. if the interface address changes
without the PIO actually changing:

	$ kill odhcpd
	$ ip addr del fd00:bbbb::1/48 dev foo-server
	$ ip addr add fd00:bbbb::2/48 dev foo-server
	$ start odhcpd
	...
	rfc9096: foo-server: changed fd00:bbbb::1/48 -> fd00:bbbb::2/48

But the PIO hasn't actually changed.

Fix this by only storing the actual prefix and length that have been
announced, i.e. fd00:bbbb::/64 and not fd00:bbbb::1/48. This also means
that the data in the JSON file is clearer, since the prefix will match
what is actually announced via the RAs.

Signed-off-by: David Härdeman <david@hardeman.nu>