router: improve comparison in router_clear_duplicated_ra_pio()
Note that time_t will have an alignment of 4 (unlikely) or 8, meaning
that struct ra_pio will have padding. A straight memcmp() could
therefore generate the wrong result unless both struct ra_pios that are
compared are guaranteed to have been memset to zero.

Avoid the risk of something going wrong here by doing a straight
comparison of the two members of the struct that we want to compare.

Signed-off-by: David Härdeman <david@hardeman.nu>