Revert "router: optimize duplicated PIO comparison"
This reverts commit d354ebb66cdcccc3373156468eec1c660fb2922c.

As stated by David Härdeman at [1]:
Note that time_t will have an alignment of 4 (unlikely) or 8, meaning
that struct ra_pio will have padding. A straight memcmp() could
therefore generate the wrong result unless both struct ra_pios that are
compared are guaranteed to have been memset to zero.

Avoid the risk of something going wrong here by doing a straight
comparison of the two members of the struct that we want to compare.

[1]: https://github.com/openwrt/odhcpd/pull/348/changes/d02fce006abd4f0761bf3100c9ddd4ce70847edb

Link: https://github.com/openwrt/odhcpd/pull/354
Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>