config: allow minimum PD len up to 64
The current limit of [1,62] doesn't really seem to be justified. With
things like RFC9762 (DHCPv6-PD preferred), it isn't unlikely that we'll
see more DHCPv6-PD requests in the future, and not just from actual
routers.

The smallest realistic PD that can be allocated is a /64, so make that
the minimum.

Also, fix a bug in the logic (note that the PD_MIN_LEN_MAX actually
isn't enforced).

Closes: #352

(cherry picked from commit 62113d007a6a18668cdcbb52811a1b79630f5848)
Signed-off-by: David Härdeman <david@hardeman.nu>
Link: https://github.com/openwrt/odhcpd/pull/357
Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>