config: add default value for dhcpv6_pd_min_len
Right now, dhcpv6_pd_min_len is zero, meaning any request would be
accepted (as long as it fits what we have available).

That means that with a /56 (a common residential scenario), a downstream
router can request a /57 and gobble up 1/2 of the available address
space. Even a default dhcpv6_pd_min_len value of 60 means that we'd
support 15 PDs in total.

So let's err on the side of caution and set /62 as the default, people
who want larger allocations can easily override the default.

Signed-off-by: David Härdeman <david@hardeman.nu>
Link: https://github.com/openwrt/odhcpd/pull/357
Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>