odhcpd: use struct ether_addr to store MAC address
Using a struct ether_addr means that we can skip specifying a length in
a number of places, makes it possible to replace the homegrown
odhcpd_print_mac() with libc equivalents (ether_ntoa()), and makes a
bunch of functions shorter.

The DHCPv4 server anyway only supports MAC addresses (and always has),
so this also makes that more explicit.

Signed-off-by: David Härdeman <david@hardeman.nu>