odhcpd: rename some leasefile functions/variables
Mea culpa, I intentionally named this "*state*" because I was planning
to extend the existing "leasefile" to store all kinds of state.

I now plan to introduce the "proper" statefile along the old leasefile
until there are no users of the leasefile, and then remove the
leasefile, so rename the existing functionality in preparation for the
following patches.

Signed-off-by: David Härdeman <david@hardeman.nu>